refactor: move authui and config
